requirements.txt
tensorflow-gpu == 2.2.0
tensorflow-probability == 0.10.0
torch # for data processing
torchaudio # for data processing
numpy